# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface Book {
  id: number;
  title: string;
  author?: string | null;
  genre?: string | null;
  totalPages: number;
  pagesRead: number;
  /** YYYY-MM-DD */
  targetCompletionDate?: string | null;
  completedAt?: string | null;
  createdAt: string;
}

export interface Session {
  id: number;
  bookId: number;
  bookTitle: string;
  pagesRead: number;
  notes?: string | null;
  loggedAt: string;
}

export interface BookDetail {
  book: Book;
  sessions: Session[];
}

export interface NewBookInput {
  title: string;
  author?: string | null;
  genre?: string | null;
  /** @minimum 1 */
  totalPages: number;
  targetCompletionDate?: string | null;
}

export interface BookUpdateInput {
  title?: string;
  author?: string | null;
  genre?: string | null;
  /** @minimum 1 */
  totalPages?: number;
  targetCompletionDate?: string | null;
}

export interface NewSessionInput {
  bookId: number;
  /** @minimum 1 */
  pagesRead: number;
  notes?: string | null;
}

export interface GoalRow {
  bookId: number;
  bookTitle: string;
  pagesCompleted: number;
  totalPages: number;
  estimatedDateCompleted?: string | null;
  avgPagesPerDay: number;
}

export interface StatsSummary {
  totalPagesRead: number;
  booksCompleted: number;
  booksInProgress: number;
  sessionsLogged: number;
  pagesToday: number;
  pagesThisWeek: number;
  avgPagesPerDay: number;
}

export interface DailyStat {
  date: string;
  pages: number;
}

export interface StreakResponse {
  currentStreak: number;
  longestStreak: number;
}

export type ListBooksParams = {
  status?: ListBooksStatus;
};

export type ListBooksStatus =
  (typeof ListBooksStatus)[keyof typeof ListBooksStatus];

export const ListBooksStatus = {
  reading: "reading",
  completed: "completed",
  all: "all",
} as const;

export type ListSessionsParams = {
  limit?: number;
  bookId?: number;
};

export type GetDailyStatsParams = {
  days?: number;
};