# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  id: number;
  title: string;
  author?: string | null;
  genre?: string | null;
  totalPages: number;
  pagesRead: number;
  /** YYYY-MM-DD */
  targetCompletionDate?: string | null;
  completedAt?: string | null;
  createdAt: string;
}