# lib/api-zod/src/generated/types/bookDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";
import type { Session } from "./session";

export interface BookDetail {
  book: Book;
  sessions: Session[];
}