# lib/api-zod/src/generated/types/goalRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalRow {
  bookId: number;
  bookTitle: string;
  pagesCompleted: number;
  totalPages: number;
  estimatedDateCompleted?: string | null;
  avgPagesPerDay: number;
}