# lib/api-zod/src/generated/types/newBookInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface NewBookInput {
  title: string;
  author?: string | null;
  genre?: string | null;
  /** @minimum 1 */
  totalPages: number;
  targetCompletionDate?: string | null;
}