# lib/api-zod/src/generated/types/newSessionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface NewSessionInput {
  bookId: number;
  /** @minimum 1 */
  pagesRead: number;
  notes?: string | null;
}