# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface Session {
  id: number;
  bookId: number;
  bookTitle: string;
  pagesRead: number;
  notes?: string | null;
  loggedAt: string;
}