# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalPagesRead: number;
  booksCompleted: number;
  booksInProgress: number;
  sessionsLogged: number;
  pagesToday: number;
  pagesThisWeek: number;
  avgPagesPerDay: number;
}